add new config method to config tree factory

+++ core/src/main/java/imgb/rulesystem/factory/streamfactory/sorter/RuleSorter.java
@@ -35,7 +35,7 @@ public class RuleSorter {
             logger.warn("rule node is null, quit");
             return;
         }
-        Integer priority = priorityManager.getPriority(rootNode);
+        Float priority = priorityManager.getPriority(rootNode);
         ruleList.add(new RuleUnit(priority, rootNode));
     }
 
@@ -61,15 +61,15 @@ public class RuleSorter {
 
     class RuleUnit implements Comparable<RuleUnit> {
 
-        private int priority;
+        private float priority;
         private BaseNode node;
 
-        public RuleUnit(int priority, BaseNode node) {
+        public RuleUnit(float priority, BaseNode node) {
             this.priority = priority;
             this.node = node;
         }
 
-        public int getPriority() {
+        public float getPriority() {
             return priority;
         }
 
@@ -78,7 +78,7 @@ public class RuleSorter {
         }
 
         public int compareTo(RuleUnit unit) {
-            return unit.getPriority() - priority;
+            return (int)(unit.getPriority() - priority);
         }
     }
 }
